+++ bgstally/activity.py
@@ -246,6 +246,9 @@ class Activity:
                     faction = system['Factions'].get(effect_faction_name)
                     if not faction: continue
 
+                    # Show activity indicator
+                    self.bgstally.ui.indicate_activity = True
+
                     if inftrend == "UpGood" or inftrend == "DownGood":
                         if effect_faction_name == journal_entry['Faction']:
                             faction['MissionPoints'] += inf
@@ -268,6 +271,8 @@ class Activity:
                     or (faction['FactionState'] in STATES_WAR and journal_entry['Name'] in MISSIONS_WAR) \
                     and effect_faction_name == journal_entry['Faction']:
                         faction['MissionPoints'] += 1
+                        # Show activity indicator
+                        self.bgstally.ui.indicate_activity = True
 
         # Thargoid War
         if journal_entry['Name'] in MISSIONS_TW_COLLECT + MISSIONS_TW_EVAC_LOW + MISSIONS_TW_EVAC_MED + MISSIONS_TW_EVAC_HIGH + MISSIONS_TW_MASSACRE and mission is not None:
@@ -283,6 +288,8 @@ class Activity:
                         tw_stations[mission_station] = self._get_new_tw_station_data(mission_station)
 
                     if mission.get('PassengerCount', -1) > -1:
+                        # Show activity indicator
+                        self.bgstally.ui.indicate_activity = True
                         if journal_entry['Name'] in MISSIONS_TW_EVAC_LOW:
                             tw_stations[mission_station]['passengers']['l']['count'] += 1
                             tw_stations[mission_station]['passengers']['l']['sum'] += mission.get('PassengerCount', -1)
@@ -293,6 +300,8 @@ class Activity:
                             tw_stations[mission_station]['passengers']['h']['count'] += 1
                             tw_stations[mission_station]['passengers']['h']['sum'] += mission.get('PassengerCount', -1)
                     elif mission.get('CommodityCount', -1) > -1:
+                        # Show activity indicator
+                        self.bgstally.ui.indicate_activity = True
                         match journal_entry.get('Commodity'):
                             case "$OccupiedCryoPod_Name;":
                                 if journal_entry['Name'] in MISSIONS_TW_EVAC_LOW:
@@ -308,6 +317,8 @@ class Activity:
                                 tw_stations[mission_station]['cargo']['count'] += 1
                                 tw_stations[mission_station]['cargo']['sum'] += mission.get('CommodityCount', -1)
                     elif mission.get('KillCount', -1) > -1:
+                        # Show activity indicator
+                        self.bgstally.ui.indicate_activity = True
                         match journal_entry.get('TargetType'):
                             case "$MissionUtil_FactionTag_Scout;":
                                 tw_stations[mission_station]['massacre']['s']['count'] += 1
@@ -340,6 +351,9 @@ class Activity:
         if mission is None: return
         self.dirty = True
 
+        # Show activity indicator
+        self.bgstally.ui.indicate_activity = True
+
         for system in self.systems.values():
             if mission['System'] != system['System']: continue
 
@@ -359,6 +373,9 @@ class Activity:
         if not current_system: return
         self.dirty = True
 
+        # Show activity indicator
+        self.bgstally.ui.indicate_activity = True
+
         faction = current_system['Factions'].get(state.station_faction)
         if faction:
             faction['CartData'] += journal_entry['TotalEarnings']
@@ -372,6 +389,9 @@ class Activity:
         current_system = self.systems.get(state.current_system_id)
         if not current_system: return
         self.dirty = True
+
+        # Show activity indicator
+        self.bgstally.ui.indicate_activity = True
 
         faction = current_system['Factions'].get(state.station_faction)
         if faction:
@@ -387,6 +407,9 @@ class Activity:
         current_system = self.systems.get(state.current_system_id)
         if not current_system: return
         self.dirty = True
+
+        # Show activity indicator
+        self.bgstally.ui.indicate_activity = True
 
         for bv_info in journal_entry['Factions']:
             faction = current_system['Factions'].get(bv_info['Faction'])
@@ -406,6 +429,9 @@ class Activity:
         if not current_system: return
         self.dirty = True
 
+        # Show activity indicator
+        self.bgstally.ui.indicate_activity = True
+
         faction = current_system['Factions'].get(journal_entry['Faction'])
         if faction:
             faction['CombatBonds'] += journal_entry['Amount']
@@ -423,6 +449,9 @@ class Activity:
         if faction:
             self.dirty = True
             bracket:int = 0
+
+            # Show activity indicator
+            self.bgstally.ui.indicate_activity = True
 
             if self.bgstally.market.available(journal_entry['MarketID']):
                 market_data:dict = self.bgstally.market.get_commodity(journal_entry['Type'])
@@ -447,6 +476,9 @@ class Activity:
             cost:int = journal_entry['Count'] * journal_entry['AvgPricePaid']
             profit:int = journal_entry['TotalSale'] - cost
             bracket:int = 0
+
+            # Show activity indicator
+            self.bgstally.ui.indicate_activity = True
 
             if journal_entry.get('BlackMarket', False):
                 faction['BlackMarketProfit'] += profit
@@ -492,12 +524,18 @@ class Activity:
                     faction['Murdered'] += 1
                     self.recalculate_zero_activity()
 
+                    # Show activity indicator
+                    self.bgstally.ui.indicate_activity = True
+
             case 'onFoot_murder':
                 # For on-foot murders, get the faction from the journal entry
                 faction = current_system['Factions'].get(journal_entry['Faction'])
                 if faction:
                     faction['GroundMurdered'] += 1
                     self.recalculate_zero_activity()
+
+                    # Show activity indicator
+                    self.bgstally.ui.indicate_activity = True
 
 
     def settlement_approached(self, journal_entry: Dict, state:State):
@@ -518,12 +556,19 @@ class Activity:
         if journal_entry.get('VictimFaction') == "$faction_Thargoid;":
             tw_ship:str = TW_CBS.get(journal_entry['Reward'])
             if tw_ship: current_system['TWKills'][tw_ship] += 1
+
+            # Show activity indicator
+            self.bgstally.ui.indicate_activity = True
+
             return
 
         # Otherwise, must be on-ground for CB kill tracking
         if state.last_settlement_approached == {}: return
 
         self.dirty = True
+
+        # Show activity indicator
+        self.bgstally.ui.indicate_activity = True
 
         timedifference = datetime.strptime(journal_entry['timestamp'], "%Y-%m-%dT%H:%M:%SZ") - datetime.strptime(state.last_settlement_approached['timestamp'], "%Y-%m-%dT%H:%M:%SZ")
         if timedifference > timedelta(minutes=5):
@@ -635,6 +680,9 @@ class Activity:
                 system['TWSandR'][key]['delivered'] += allocatable
                 count -= allocatable
                 self.dirty = True
+
+                # Show activity indicator
+                self.bgstally.ui.indicate_activity = True
 
         # count can end up > 0 here - i.e. more S&R handed in than we originally logged as scooped. Ignore, as we don't know
         # where it originally came from