+++ bgstally/ui.py
@@ -44,9 +44,7 @@ class UI:
         self.image_button_cmdrs = PhotoImage(file = path.join(self.bgstally.plugin_dir, FOLDER_ASSETS, "button_cmdrs.png"))
         self.image_button_carrier = PhotoImage(file = path.join(self.bgstally.plugin_dir, FOLDER_ASSETS, "button_carrier.png"))
 
-        self.thread: Optional[Thread] = Thread(target=self._worker, name="BGSTally UI worker")
-        self.thread.daemon = True
-        self.thread.start()
+        self.indicate_activity:bool = False
 
         # Single-instance windows
         self.window_cmdrs:WindowCMDRs = WindowCMDRs(self.bgstally)
@@ -54,6 +52,10 @@ class UI:
         self.window_legend:WindowLegend = WindowLegend(self.bgstally)
         # TODO: When we support multiple APIs, this will no longer be a single instance window
         self.window_api:WindowAPI = WindowAPI(self.bgstally, self.bgstally.api_manager.apis[0])
+
+        self.thread: Optional[Thread] = Thread(target=self._worker, name="BGSTally UI worker")
+        self.thread.daemon = True
+        self.thread.start()
 
 
     def shut_down(self):
@@ -186,6 +188,12 @@ class UI:
             self.bgstally.overlay.display_message("tick", f"Curr Tick: {self.bgstally.tick.get_formatted(DATETIME_FORMAT_OVERLAY)}", True)
             minutes_delta:int = int((datetime.utcnow() - self.bgstally.tick.next_predicted()) / timedelta(minutes=1))
 
+            # Activity Indicator
+            if self.indicate_activity:
+                self.bgstally.overlay.display_message("info", " ", True, 2)
+                self.indicate_activity = False
+
+            # Tick Warning
             if datetime.utcnow() > self.bgstally.tick.next_predicted() + timedelta(minutes = TIME_TICK_ALERT_M):
                 self.bgstally.overlay.display_message("tickwarn", f"Tick {minutes_delta}m Overdue (Estimated)", True)
             elif datetime.utcnow() > self.bgstally.tick.next_predicted():